Add snapshot test for Sidebar and move to components
