+++ src/components/Sidebar/index.js
@@ -5,6 +5,7 @@ export default class Sidebar extends Component {
   render() {
     return (
       <aside className="Sidebar">
+        <h1>Paired</h1>
         <NavLink to="/schedule">View Schedule</NavLink>
         <NavLink to="/book-pairing">Book a Pairing</NavLink>
       </aside>
